+++ xrGame/game_base.cpp
@@ -145,7 +145,7 @@ bool game_PlayerState::HasOldID(u16 ID)
 
 game_GameState::game_GameState()
 {
-	m_type = -1;
+	m_type = static_cast<u32>(-1);
 	m_phase = GAME_PHASE_NONE;
 	m_round = -1;
 	m_round_start_time_str[0] = 0;
